removing warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5344 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSLane.h
@@ -230,7 +230,7 @@ struct VehPosition : public std::binary_function< const MSVehicle*,
     /// returns the vehicles
     virtual const VehCont &getVehiclesSecure();
 
-    size_t getVehicleNumber() const;
+    unsigned int getVehicleNumber() const;
